requirements copy.txt
langgraph==0.4.1
langchain==0.3.25
voyageai==0.3.2
python-dotenv==1.1.0
langchain-groq==0.3.2
langfuse==2.59.4
qdrant-client==1.14.2
pymupdf
boto3==1.26.160 
botocore==1.29.160
psycopg2==2.9.10
sentence-transformers==5.1.0
chromadb==1.1.0
PyJWT==2.10.1